==> 1) Mapped_ids_Matching.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing
import os

# Merge two mapid dataset and save mapped_id data
mapped_data_threshold = pd.read_csv('.../mapped_ids_threshold.csv')
mapped_data_worsening = pd.read_csv('.../mapped_ids_worsening.csv')

mapped_data = pd.concat([mapped_data_threshold, mapped_data_worsening])
mapped_data = mapped_data[["hashed_patient_id","patient_id"]]
mapped_data = mapped_data.drop_duplicates(subset="hashed_patient_id")
mapped_data.to_csv("mapped_id.csv", index = False) # 1069

# Adding "patient id" to air pollution data by merging airpollution data and mapped id data
Airquality_data = pd.read_csv(".../MS_patients_meas_filt.csv")

Airquality = pd.merge(mapped_data,Airquality_data, 
                      how = "inner", left_on='hashed_patient_id',
                      right_on='patient_id')
Airquality.drop(["patient_id_y"], axis = 1, inplace = True)
Airquality.rename(columns={"patient_id_x": "patient_id"}, inplace = True)

# Create an empty list to store matching values
matching_values = []

# Iterate through the rows of the first dataset
for index, row in Airquality_data.iterrows():
    patient_id = row["patient_id"]
    
    # Check if the patient_id exists in the second dataset's hashed_patient_id column
    if patient_id in mapped_data_threshold["hashed_patient_id"].values:
        matching_values.append(patient_id)

# Create a DataFrame from the list of matching values
matching_data = pd.DataFrame({"patient_id": matching_values})

# Display the result
print(matching_data)


Airquality.to_csv("Updated_Airquality.csv", index = False) # 814
