==> 4) WHO_Percentage_Calculation.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing
from scipy.stats import linregress
import datetime

Clinical_data_as_2013 = pd.read_csv("...//Updated_Clinical_Data_Since2013.csv")
CAP_mondino = pd.read_csv("...//Patient_CAP.csv")
IDW_data = pd.read_csv(".../weighted_measurement_PV.csv")
IDW_data.head()

# This celle try to select air pollution data between visit to visit
Clinical_data_with_CAP['Date_of_visit'] = pd.to_datetime(Clinical_data_with_CAP['Date_of_visit'], 
                                                         format ='%Y-%m-%d')

IDW_data["timestamp"] = pd.to_datetime(IDW_data["timestamp"],format ='%Y-%m-%d')

# WHO Percentage features for visit to visit time window
Threshold_values ={
    "PM25" : 15,
    "PM10" : 45,
    "N02" : 25,
    "O3" : 100,
    "SO2" : 40,
    "CO" : 4}

matched_data = []
for index, row in Clinical_data_with_CAP.iterrows():
    print(index)
    patient_postcode = row['CAP']
    patient_id = row['Patient_ID']
    visit_date = row['Date_of_visit']
    if len(Clinical_data_with_CAP) == index+1:
        break
    elif Clinical_data_with_CAP.iloc[index+1]["Patient_ID"] == patient_id:
        next_visit_date = Clinical_data_with_CAP.loc[index+1,'Date_of_visit']
        
        if patient_postcode in list(IDW_data["postcode_pat"]):
            
            matching_rows = IDW_data[(IDW_data['postcode_pat'] == patient_postcode) &
                             (IDW_data['timestamp'].dt.date >= visit_date.date()) &
                             (IDW_data['timestamp'].dt.date < next_visit_date.date())]

            
            if len(matching_rows) != 0:
                
                for name in Threshold_values.keys():
                    air_quality = matching_rows[matching_rows["name"] == name]
                    
                    if len(air_quality) != 0:
                        over_threshold_data  = air_quality[air_quality["measurement"] >
                                                          Threshold_values[name]]
                        Clinical_data_with_CAP.loc[index+1,"WHOPercent_"+ name] = round((len(over_threshold_data) / 
                                                                                         len(air_quality)), 3)
                    else:
                        continue
            else:
                print("matching rows is zero")
        else:
            print("patient postcode not found " + str(patient_postcode))
    else:
        continue 

Clinical_data_with_CAP.to_csv("Data_with_WHO.csv", index= False)

for CAP in Clinical_data_with_CAP["CAP"].unique():
    CAP_data = Clinical_data_with_CAP[Clinical_data_with_CAP["CAP"]==CAP]
    for patient_id in CAP_data["Patient_id"].unique():
        patient_data = CAP_data[CAP_data["Patient_id"]==patient_id]
        for row in range(len(patient_data)-1):
            prev_row = patient_data.iloc[row]
            next_row = patient_data.iloc[row+1]
            env_prev_row = Env_data[(Env_data["week_from_baseline"] == prev_row["week_from_baseline"])&
                                (Env_data["patient_id"] == prev_row["patient_id"])]
            env_next_row = Env_data[(Env_data["week_from_baseline"] == next_row["week_from_baseline"])&
                                (Env_data["patient_id"] == next_row["patient_id"])]
            mean_values = Env_data.iloc[env_prev_row.index[0]:env_next_row.index[0]].mean()["PM25_mean":"SO2_who_percentage"]
            FS_Score.loc[next_row.reset_index().columns[1],mean_values.index] = mean_values
