==> 8) Inverse_Distance_Weighting-4_Weeks_Ago.ipynb <==
import numpy as np
import pandas as pd
import math
from tqdm import tqdm

# Measuring Distance between Patients and Stations

#Plat = Patient Latitude
#Plon = Patient Longitude
#Slat = Station Latitude
#Slon = Station Longitude

def earth_dist(Plat,Plon,Slat,Slon):
    

    rad = math.pi / 180
    
    a1 = Plat * rad
    a2 = Plon * rad
    b1 = Slat * rad
    b2 = Slon * rad
    
    distance_lat = b1 - a1
    distance_lon = b2 - a2
    
    a = math.sin(distance_lat/2)**2 + math.cos(a1) * math.cos(b1) * math.sin(distance_lon/2)**2
    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))
    R = 6371.0  # Radius of the Earth in kilometers
    d = R * c
    
    return d

def inverse_distance_weighting(distances, measurements, power=2):
    
    numerator = np.sum(measurements / (distances ** power))
    weights = np.sum(1 / (distances ** power))
    weighted_measurement = numerator / weights
    
    return weighted_measurement

#Station
Path_Station = ".../postcode_of_station.csv"
postcode_station = pd.read_csv(Path_Station)
postcode_station = postcode_station[postcode_station["country"]==2]

#Patient
Path_Patient = ".../postcode_of_patient.csv"
postcode_patient = pd.read_csv(Path_Patient)
postcode_patient = postcode_patient[postcode_patient["country"]==2]

#Measurement
Path_Measurement = ".../measurement_on_postcode_of_station.csv"
airpollution_measurement = pd.read_csv(Path_Measurement)

#Updated FS-Score & Airpollution
Path_4weeks = ".../Imputed_4Weeks_Final_Data.csv"
FS_Pollutant_data = pd.read_csv(Path_4weeks)


# Finding n nearest stations for each patient

FS_Pollutant_data['Start_date'] = pd.to_datetime(FS_Pollutant_data['Start_date'])
FS_Pollutant_data["End_date"] = pd.to_datetime(FS_Pollutant_data["End_date"])
FS_Pollutant_data['Start_date_4WeeksAgo'] = pd.to_datetime(FS_Pollutant_data['Start_date_4WeeksAgo'])
FS_Pollutant_data["end_data_4WeeksAgo"] = pd.to_datetime(FS_Pollutant_data["end_data_4WeeksAgo"])
FS_Pollutant_data["Date_of_visit"] = pd.to_datetime(FS_Pollutant_data["Date_of_visit"])


FS_Pollutant_data["Station_postcode"] = FS_Pollutant_data["Station_postcode"].astype("str")
FS_Pollutant_data = pd.merge(postcode_station, FS_Pollutant_data, 
                                   left_on='postcode', right_on = "Station_postcode", 
                                   how='inner')
print(FS_Pollutant_data.shape)
FS_Pollutant_data = FS_Pollutant_data[['Patient_id', 'Patient_postcode', 'Station_postcode',
                                       'id_postcode_of_station', 'latitude','longitude',
                                       'Date_of_visit', 'Start_date', 'End_date', 'Start_date_4WeeksAgo',
                                        'end_data_4WeeksAgo', 'EDSS_score_assessed_by_clinician', 'Sex', 'Age',
                                        'MS.in.pediatric.age', 'Relapse_number', 'Pyramidal', 'Cerebellar',
                                        'Thronchioencephalic', 'Sensitive', 'Sphincteric', 'Visual', 'Mental',
                                        'Deambulation', 'PM25_mean', 'PM10_mean', 'CO_mean', 'NO2_mean',
                                        'O3_mean', 'SO2_mean', 'FG_mean', 'HU_mean', 'PP_mean', 'QQ_mean',
                                        'RR_mean', 'TG_mean', 'TN_mean', 'TX_mean']]

FS_Pollutant_data.rename(columns={"latitude": "Station_lat", "longitude": "Station_long"},inplace = True)
FS_Pollutant_data.to_csv("FS_Pollutant_data.csv", index = False)

radius = 50 # kilometer
count = 0
result = []
for index_p in FS_Pollutant_data["Station_postcode"].unique():
    row_p = FS_Pollutant_data[FS_Pollutant_data["Station_postcode"]==index_p].iloc[0]
    temp = []
    for index_s, row_s in postcode_station.iterrows():
        dist = earth_dist(row_p["Station_lat"],row_p["Station_long"],
                          row_s["latitude"],row_s["longitude"])
        
        temp.append([dist, 
                    row_s["id_postcode_of_station"]])
        
    sorted_list = sorted(temp, key=lambda x: x[0])
    n_nearest_stations = [x for x in sorted_list if x[0] <= radius]
    if not n_nearest_stations:
        print("no station within 50 kilometer for postcode number " + row_p["Station_postcode"])
        count = count + 1
    else:
        result.append([row_p["Station_postcode"],
                       n_nearest_stations])

data = []                     
for s_postcode, nearest_stations in result:
    for dist, s_id in nearest_stations:
        data.append([s_postcode,
                     dist,
                     s_id,
                     len(nearest_stations)])

nearest_stations_to_patients = pd.DataFrame(data, columns=["Station_postcode",
                                                            "Distance(km)",
                                                            "Nearest_station_postcode",
                                                            "Station_Number"])

nearest_stations_to_patients.to_csv("Nearest_Station_Locations.csv", index = False)
len(nearest_stations_to_patients["Station_postcode"].unique()) # 131
nearest_stations_to_patients.head() # 3801*5

# Your existing code to define column_to_param
FS_Pollutant_data['Start_date'] = pd.to_datetime(FS_Pollutant_data['Start_date'])
FS_Pollutant_data["End_date"] = pd.to_datetime(FS_Pollutant_data["End_date"])
FS_Pollutant_data['Start_date_4WeeksAgo'] = pd.to_datetime(FS_Pollutant_data['Start_date_4WeeksAgo'])
FS_Pollutant_data["end_data_4WeeksAgo"] = pd.to_datetime(FS_Pollutant_data["end_data_4WeeksAgo"])
FS_Pollutant_data["Date_of_visit"] = pd.to_datetime(FS_Pollutant_data["Date_of_visit"])
column_to_param = {
    'PM10_mean':1,
    'PM25_mean':2,
    'SO2_mean':3,
    'O3_mean':5,
    'NO2_mean':6,
    'CO_mean':7,
    'FG_mean':8,
    'HU_mean':9, 
    'PP_mean':10, 
    'QQ_mean':11,
    'RR_mean':12, 
    'TG_mean':13, 
    'TN_mean':14, 
    'TX_mean':15  
}

airpollution_measurement['timestamp'] = pd.to_datetime(airpollution_measurement['timestamp'])

for index, row in tqdm(FS_Pollutant_data.iterrows()):
    print(index)
    missing_columns = row[row.isnull()].index.tolist()
    if missing_columns:
        row_postcode = row["Station_postcode"]
        date_of_visit = row["Date_of_visit"]
        nearest_stations = nearest_stations_to_patients[nearest_stations_to_patients["Station_postcode"]==row_postcode]
        for param in missing_columns:
            
            print(param)
            # Filter rows based on postcode, timestamp, and param
            filtered_rows = airpollution_measurement[
                (airpollution_measurement['postcode'].isin(nearest_stations["Nearest_station_postcode"]))
                & (airpollution_measurement['timestamp'].dt.date == date_of_visit.date())
                & (airpollution_measurement['param'] == column_to_param[param])]
            
            if filtered_rows.empty:
                
                print("No measurement for all nearest neighbors on {} for postcode number {} for param {}"
                      .format(date_of_visit, row_postcode, param))
            else:
                
                exist_postcode = nearest_stations[nearest_stations["Nearest_station_postcode"].isin(filtered_rows["postcode"])]

                measurment_nearest = filtered_rows.merge(exist_postcode, 
                                                         left_on='postcode', 
                                                         right_on='Nearest_station_postcode')

                measurment_nearest = measurment_nearest[['Station_postcode', 'Nearest_station_postcode',
                                                         'Distance(km)','measurement', 
                                                         'timestamp', 'param','Station_Number']]

                distances = measurment_nearest["Distance(km)"]
                measurements = measurment_nearest["measurement"]
                
                weighted_measurement = inverse_distance_weighting(distances, measurements, power=2)
                row[param] = round(weighted_measurement,2)
                FS_Pollutant_data.iloc[index] = row

FS_Pollutant_data.to_csv("Imputed_4Weeks_Merged_Final_data_1.csv", index = False)

#iloc: 0 to2608
Path_Data_1 = ".../Imputed_4Weeks_Merged_AirFS_2608.csv"
Data_1 = pd.read_csv(Path_Data_1)

#iloc: 2608 to 3347
Path_Data_2 = ".../Imputed_4Weeks_Merged_AirFS3347.csv"
Data_2 = pd.read_csv(Path_Data_2)

#iloc: 3347 to end
Path_Data_3 = ".../Imputed_4Weeks_Merged_AirFS_end.csv"
Data_3 = pd.read_csv(Path_Data_3)


Data_3['Start_date'] = pd.to_datetime(Data_3['Start_date'])
Data_3["End_date"] = pd.to_datetime(Data_3["End_date"])
Data_3['Start_date_4WeeksAgo'] = pd.to_datetime(Data_3['Start_date_4WeeksAgo'])
Data_3["end_data_4WeeksAgo"] = pd.to_datetime(Data_3["end_data_4WeeksAgo"])
Data_3["Date_of_visit"] = pd.to_datetime(Data_3["Date_of_visit"])
result = pd.concat([Data_1.iloc[0:2608], Data_2.iloc[2608:3374], Data_3.iloc[3374:]])
result.to_csv("Imputed_4Weeks_Final_Data.csv", index = False)
