==> 9) Complete_Case_Air_FS.ipynb <==
import seaborn as sns
import pandas as pd
import numpy as np

def calculate_age_at_visit(row):
    date_of_birth = row['Data di nascita']
    date_of_visit = row['Date_of_visit']
    age = date_of_visit.year - date_of_birth.year - ((date_of_visit.month, date_of_visit.day) < (date_of_birth.month, date_of_birth.day))
    return age

def count_relapses(patient_id, date):
    
    relapses = Relapse_number[(Relapse_number['Paziente ID'] == patient_id)
                            & (Relapse_number['relapse'] <= date)]
    
    if len(relapses) < 1:
        return 0
    else:
        return relapses.iloc[-1]['Relapse Count']

path = ".../necessary_data.csv"
Air_data =pd.read_csv(path)
Air_data.head(6)

path = ".../Updated_MS.csv"
FS_data =pd.read_csv(path)

Relapse_number = pd.read_csv(".../relapse.csv")

FS_data['Date_of_visit'] = pd.to_datetime(FS_data['Date_of_visit'], format ='%d/%m/%Y')
FS_data['Date of birth'] = pd.to_datetime(FS_data['Date of birth'], format ='%d/%m/%Y')
Air_data['start_date'] = pd.to_datetime(Air_data['start_date'], format ='%d/%m/%Y')
Air_data['end_date'] = pd.to_datetime(Air_data['end_date'], format ='%d/%m/%Y')
FS_data['Patient_ID']=FS_data['Patient_ID'].astype(object)
Relapse_number['relapse'] = pd.to_datetime(Relapse_number['relapse'])
FS_data['Relapse_number'] = FS_data.apply(lambda row:count_relapses(row['Patient_ID'],
                                                                    row['Date_of_visit']),
                                          axis=1)

FS_Air_data = FS_data.merge(Air_data, left_on='Patient_ID', right_on = "patient_id", how='inner')
Final_data = FS_Air_data[(FS_Air_data['start_date'] <= FS_Air_data['Date_of_visit']) &
            (FS_Air_data['end_date'] >= FS_Air_data['Date_of_visit'])]
Final_data.dropna(inplace = True)

Final_data = Final_data[['Patient_ID', 'Date_of_visit','start_date', 'end_date',
                         'Pyramidal', 'Cerebellar','Thronchioencephalic', 'Sensitive',
                         'Sphincteric', 'Visual', 'Mental','Deambulation', 
                         'EDSS_score_assessed_by_clinician', 'Date of birth','Sex',
                         'MS in pediatric age', 'Relapse_number','PM25_mean','PM10_mean',
                         'CO_mean', 'NO2_mean', 'O3_mean', 'SO2_mean','FG_mean', 'HU_mean',
                         'PP_mean', 'QQ_mean', 'RR_mean', 'TG_mean','TN_mean', 'TX_mean' ]]
Final_data.head()

Final_data.to_csv("FS_Pollutant_Completed.csv", index= False)
