==> 10) Shift_Code.ipynb <==
import numpy as np
import pandas as pd
import math
import matplotlib.pyplot as plt
from scipy.stats import linregress
import datetime

# Necessary Datasets
FS_Scores_data = pd.read_csv('.../Inverse Distance weaghting/Final_data_Description.csv')
FS_Scores_data.head(20)

# Doing some preprocessing for both data
# Remove duplicate rows in clinical data
features_to_shift = FS_Scores_data.columns[15:]

# Shift the values of the selected features to the previous visit
for feature in features_to_shift:
    FS_Scores_data[feature] = FS_Scores_data.groupby('Patient_id')[feature].shift(-1)

FS_Scores_data[features_to_shift] = FS_Scores_data.groupby('Patient_id')[features_to_shift].ffill()

FS_Scores_data.to_csv("Shifted_data.csv", index= False)
